gateway-service/pyproject.toml
[project]
name = "eccomerce-gateway-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
